+++ lib/Perl/Tidy/IOScalar.pm
@@ -10,7 +10,7 @@ package Perl::Tidy::IOScalar;
 use strict;
 use warnings;
 use Carp;
-our $VERSION = '20220613.03';
+our $VERSION = '20220613.04';
 
 use constant EMPTY_STRING => q{};