+++ lib/Perl/Tidy/LineSink.pm
@@ -8,7 +8,7 @@
 package Perl::Tidy::LineSink;
 use strict;
 use warnings;
-our $VERSION = '20220613.03';
+our $VERSION = '20220613.04';
 
 sub AUTOLOAD {